refactor resource arrivals
