+++ client/apps/game/src/hooks/helpers/use-resource-arrivals.tsx
@@ -1,24 +1,16 @@
 import { useDojo } from "@/hooks/context/dojo-context";
 import useNextBlockTimestamp from "@/hooks/use-next-block-timestamp";
-import { ContractAddress, ID, Position } from "@bibliothecadao/eternum";
+import {
+  ArrivalInfo,
+  ContractAddress,
+  DONKEY_RESOURCE_TRACKER,
+  LORDS_AND_DONKEY_RESOURCE_TRACKER,
+  LORDS_RESOURCE_TRACKER,
+} from "@bibliothecadao/eternum";
 import { useEntityQuery } from "@dojoengine/react";
 import { Entity, Has, HasValue, NotValue, defineQuery, getComponentValue, isComponentUpdate } from "@dojoengine/recs";
 import { getEntityIdFromKeys } from "@dojoengine/utils";
 import { useCallback, useEffect, useMemo, useState } from "react";
-
-const DONKEY_RESOURCE_TRACKER = 452312848583266388373324160190187140051835877600158453279131187530910662656n;
-const LORDS_RESOURCE_TRACKER = 7237005577332262213973186563042994240829374041602535252466099000494570602496n;
-const LORDS_AND_DONKEY_RESOURCE_TRACKER = 7689318425915528602346510723233181380881209919202693705745230188025481265152n;
-
-export type ArrivalInfo = {
-  entityId: ID;
-  recipientEntityId: ID;
-  position: Position;
-  arrivesAt: bigint;
-  isOwner: boolean;
-  hasResources: boolean;
-  isHome: boolean;
-};
 
 const getCurrentDonkeyWeightMinimum = () => {
   return Number(localStorage.getItem("WEIGHT_MINIMUM") || 0) * 1000;
