+++ client/sdk/packages/eternum/src/constants/resources.ts
@@ -1,6 +1,11 @@
 import { ResourceInputs, ResourceOutputs, Resources } from "../types";
 import { ResourcesIds } from "./index";
 
+export const DONKEY_RESOURCE_TRACKER = 452312848583266388373324160190187140051835877600158453279131187530910662656n;
+export const LORDS_RESOURCE_TRACKER = 7237005577332262213973186563042994240829374041602535252466099000494570602496n;
+export const LORDS_AND_DONKEY_RESOURCE_TRACKER =
+  7689318425915528602346510723233181380881209919202693705745230188025481265152n;
+
 export const resources: Array<Resources> = [
   {
     trait: "Stone",
